requirements.txt
fastapi
uvicorn[standard]
httpx
PyYAML
pydantic
google-genai
colorlog